Blockchain entity was removed to clean up codebase.
